Fixing the algorithm
I identified that the index of boxes were sometimes duplicating or skipping, this fix actually gets the algorithm to work much better. Occasionally the algorithm will decide not to use the 15th object,  this isn't a big problem - but may be worth looking into. Additionally I want to try implement weighted randomness for the objects. This would allow us to tinker with the probability of the random generator. In terms of our objects, I'd like to make the 0 appear less frequently!